move getBHdict module into level2 getBH
